Second xpath() approach for XML version detection
When upgrading my local environment from PHP 5.4 to PHP 5.5, suddenly the get_module_version() function failed. After some research I found that the xpath() method of SimpleXML has issues with the // syntax to find the version-element. This patch adds a workaround for detecting the version in a different way, if the original approach fails (like in my case).